Restructured cflisp-web, included missing license (and added link to it on page)
