+++ cflisp-web/web/index.html
@@ -14,6 +14,7 @@
 	<div class="page-wrapper">
 		<div id="source"></div>
 		<section class="output">
+			<a href="ace/LICENSE" id="license_link">Editor license</a>
 			<div class="settings">
 				<label>Optimisation: </label><input type="number" min="0" max="2" id="opt" value="1" />
 				<br />
@@ -29,5 +30,5 @@
 </body>
 
 </html>
-<script src="./ace.js" type="text/javascript" charset="utf-8"></script>
+<script src="../ace/ace.js" type="text/javascript" charset="utf-8"></script>
 <script type="module" src="script.js"></script>
