+++ cflisp-web/web/style.css
@@ -1,0 +1,47 @@
+body, html {
+	margin: 0;
+	font-family: 'Roboto Mono', monospace;
+}
+
+.page-wrapper {
+	display: flex;
+	width: 100%;
+	height: 100%;
+}
+
+.output {
+	box-sizing: border-box;
+	padding-top: 1rem;
+	display: flex;
+	flex-direction: column;
+	padding-left: 1.5rem;
+	padding-right: 1.5rem;
+	width: 50vw;
+	height: 100vh;
+}
+
+#source {
+	width: 50vw;
+	height: 100vh;
+	font-size: 1rem;
+}
+
+#result {
+	margin-top: 1rem;
+	box-sizing: border-box;
+	padding: 0.5rem;
+	margin-bottom: 1rem;
+	font-family: 'Roboto Mono', monospace;
+	box-sizing: border-box;
+	font-size: 0.95rem;
+	height: 100%;
+	border: solid;
+	border-width: 1px;
+	border-color: rgb(206, 206, 206);
+	resize: none;
+}
+
+#license_link {
+	font-size: 0.75rem;
+	text-align: right;
+}
